# feat: optimize alignment processing and remove vocabulary size limitations