@nikvaessen 2025-01-30
Thanks a lot for looking into this issue! I was not aware of this limitation. think it might be worth it to change the word hashing implementation. We could map each unique word a python integer instead. I'm not sure how much slower this is, maybe we only do it if the vocab size is larger than 0x10FFFF words. 

Can you share your thoughts behind throwing a `ValueError`, or do you agree that it might be better to fix the limitation?