@thevilledev 2025-01-30
Hey, thanks for the quick reply! I actually had an implementation ready for hashing, but it broke basically all the tests that rely on `chr()` in this case. Totally agree that hashing would be the right thing to do, and this is just an immediate fix. I can also work on that refactoring, but should we merge this in in the meanwhile?