@nikvaessen 2025-01-30
```
def _word2char(reference: List[List[str]], hypothesis: List[List[str]]):
    # tokenize each word into an integer
    vocabulary = set(chain(*reference, *hypothesis))

    if "" in vocabulary:
        raise ValueError(
            "Empty strings cannot be a word. "
            "Please ensure that the given transform removes empty strings."
        )

    word2int = dict(zip(vocabulary, range(len(vocabulary))))

    reference_ints = [
        [word2int[w] for w in sentence] for sentence in reference
    ]
    hypothesis_ints = [
        [word2int[w] for w in sentence] for sentence in hypothesis
    ]

    return reference_ints, hypothesis_ints
```
Seems to work fine for.

It seems a bit slower, but significantly:

```



current 
--------------------------------------------------------------------------------------------- benchmark: 4 tests ---------------------------------------------------------------------------------------------
Name (time in us)            Min                    Max                  Mean                StdDev                Median                   IQR            Outliers          OPS            Rounds  Iterations
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_speed_n1            19.1780 (1.0)          59.8240 (1.0)         20.5117 (1.0)          1.7625 (1.0)         20.3230 (1.0)          0.4490 (1.0)       251;441  48,752.6335 (1.0)       11379           1
test_speed_n10           92.6020 (4.83)        221.6960 (3.71)        97.2698 (4.74)         5.0923 (2.89)        96.6030 (4.75)         1.0650 (2.37)      204;976  10,280.6836 (0.21)       7027           1
test_speed_n100         782.6560 (40.81)     1,802.7650 (30.13)      834.0676 (40.66)      108.0104 (61.28)      812.2400 (39.97)       16.0885 (35.83)       19;50   1,198.9436 (0.02)        439           1
test_speed_n1000      7,985.9050 (416.41)   16,603.6110 (277.54)   9,299.2639 (453.36)   1,798.1539 (>1000.0)  8,550.4255 (420.73)   1,560.5880 (>1000.0)       8;7     107.5354 (0.00)        104           1
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

int mapping

---------------------------------------------------------------------------------------------- benchmark: 4 tests ----------------------------------------------------------------------------------------------
Name (time in us)            Min                    Max                   Mean                StdDev                 Median                   IQR            Outliers          OPS            Rounds  Iterations
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_speed_n1            19.9990 (1.0)          69.8270 (1.0)          24.7692 (1.0)          2.6409 (1.0)          24.9580 (1.0)          0.9015 (1.0)     1686;1891  40,372.6539 (1.0)        9912           1
test_speed_n10           94.5230 (4.73)        923.4460 (13.22)       110.9225 (4.48)        15.8243 (5.99)        112.0380 (4.49)        15.2380 (16.90)     249;115   9,015.3040 (0.22)       6126           1
test_speed_n100         825.8250 (41.29)    11,821.3680 (169.30)    1,047.6752 (42.30)      669.3954 (253.47)    1,001.9540 (40.15)      169.1627 (187.65)      11;22     954.4943 (0.02)        865           1
test_speed_n1000      8,345.1510 (417.28)   20,286.1870 (290.52)   10,408.9054 (420.24)   2,130.1998 (806.62)   10,090.4130 (404.30)   1,230.6775 (>1000.0)       7;7      96.0716 (0.00)         88           1
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------