@thevilledev 2025-01-30
Yeah my bad - I was wrapping them to string (and breaking it in the process). I missed that `rapidfuzz.distance.Levenshtein.editops()` supports generic sequences. Updated the PR based on the comments, thanks!

I took some optimisations from my broken implementation and aligned it with the integer-based maps. Benchmark results from my machine:

```
- current main branch:

------------------------------------------------------------------------------------------- benchmark: 4 tests ------------------------------------------------------------------------------------------
Name (time in us)            Min                   Max                  Mean              StdDev                Median                IQR            Outliers           OPS            Rounds  Iterations
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_speed_n1             9.5000 (1.0)         42.6670 (1.0)          9.9302 (1.0)        2.1570 (1.0)          9.6670 (1.0)       0.1250 (1.0)          8;41  100,702.4429 (1.0)         521           1
test_speed_n10           61.0830 (6.43)       194.6250 (4.56)        62.1981 (6.26)       2.2764 (1.06)        61.8330 (6.40)      0.3330 (2.66)     555;1103   16,077.6569 (0.16)      12013           1
test_speed_n100         572.9580 (60.31)      657.8750 (15.42)      577.7681 (58.18)      4.9037 (2.27)       576.7500 (59.66)     3.0410 (24.33)       97;85    1,730.7982 (0.02)       1582           1
test_speed_n1000      5,842.5000 (615.00)   9,644.5410 (226.04)   5,948.0211 (598.98)   412.8883 (191.42)   5,890.1245 (609.30)   42.0625 (336.50)       3;17      168.1231 (0.00)        160           1
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

- with proposed changes:

------------------------------------------------------------------------------------------- benchmark: 4 tests -------------------------------------------------------------------------------------------
Name (time in us)            Min                    Max                  Mean              StdDev                Median                IQR            Outliers           OPS            Rounds  Iterations
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_speed_n1             8.8330 (1.0)          29.6250 (1.0)          9.0443 (1.0)        0.5102 (1.0)          9.0000 (1.0)       0.0830 (1.0)        42;230  110,566.5630 (1.0)        4508           1
test_speed_n10           55.1670 (6.25)        189.3330 (6.39)        56.1595 (6.21)       2.2208 (4.35)        55.8750 (6.21)      0.2500 (3.01)     667;1140   17,806.4399 (0.16)      13683           1
test_speed_n100         518.7500 (58.73)    14,773.5410 (498.68)     561.5938 (62.09)    457.9989 (897.63)     539.5000 (59.94)    18.7193 (225.53)       6;96    1,780.6464 (0.02)       1757           1
test_speed_n1000      5,291.5420 (599.07)    9,306.3340 (314.14)   5,417.3031 (598.97)   476.7528 (934.39)   5,326.2500 (591.81)   31.1145 (374.87)       5;30      184.5937 (0.00)        179           1
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
``` 

I replaced editops with direct opcodes usage for faster alignment processing, with single-pass error counting when doing the chunking. This shows consistent speedups on all batch sizes, mainly 7-10% on smaller ones and 3-9% on larger ones, with lower standard deviance.